test-app/graphql/graphql/requirements.txt
Flask
Flask-GraphQL
graphene
graphene-sqlalchemy
psycopg2-binary
SQLAlchemy